requirements.txt
Django==2.2.6
django-bootstrap-form==3.4
django-filter==2.2.0
django-tables2==2.1.1
djangorestframework==3.10.3
pytz==2019.3
sqlparse==0.3.0
